# Add Multi-Architecture Support, Compression Levels, and Performance Benchmarks