--- a/Cs7z.Core/ISevenZipArchive.cs
+++ b/Cs7z.Core/ISevenZipArchive.cs
@@ -10,5 +10,6 @@ public interface ISevenZipArchive
         CancellationToken cancellationToken = default);
 
     Task CreateArchive(string archiveFilePath, string folder, CancellationToken cancellationToken = default);
+    Task CreateArchive(string archiveFilePath, string folder, CompressionLevel compressionLevel, CancellationToken cancellationToken = default);
     Task<ArchiveListResult> ListArchive(string archiveFilePath, CancellationToken cancellationToken = default);
 }