--- a/Cs7z.Core/Properties/AssemblyInfo.cs
+++ b/Cs7z.Core/Properties/AssemblyInfo.cs
@@ -1,0 +1,3 @@
+using System.Runtime.CompilerServices;
+
+[assembly: InternalsVisibleTo("Cs7z.Core.Tests")]