--- a/Cs7z.Service.IntegrationTests/GlobalUsings.cs
+++ b/Cs7z.Service.IntegrationTests/GlobalUsings.cs
@@ -1,0 +1,7 @@
+global using Xunit;
+global using System;
+global using System.Collections.Generic;
+global using System.IO;
+global using System.Linq;
+global using System.Threading;
+global using System.Threading.Tasks;