… CLAUDE.md
… Cs7z.Service.IntegrationTests/ListArchiveIntegrationTests.cs
… Cs7z.Service.IntegrationTests/PlatformSpecificTests.cs
… Cs7z.Service.IntegrationTests/README.md
… Cs7z.Service.IntegrationTests/TestArchiveHelper.cs
… Cs7z.Service.Linux/Cs7z.Service.Linux.csproj
… Cs7z.Service.Linux/LinuxSevenZipExecutableSource.cs
… Cs7z.Service.Linux/Source/Linux/7zz
… Cs7z.Service.Linux/Source/Linux/arm/7zz
… Cs7z.Service.Linux/Source/Linux/arm/7zzs
… Cs7z.Service.Linux/Source/Linux/arm64/7zz
… Cs7z.Service.Linux/Source/Linux/arm64/7zzs
… Cs7z.Service.Linux/Source/Linux/x64/7zz
… Cs7z.Service.Linux/Source/Linux/x64/7zzs
… Cs7z.Service.Linux/Source/Linux/x86/7zz
… Cs7z.Service.Linux/Source/Linux/x86/7zzs
… Cs7z.Service.MacOsx/Cs7z.Service.MacOsx.csproj
… Cs7z.Service.MacOsx/MacSevenZipExecutableSource.cs
… Cs7z.Service.MacOsx/Source/osx/arm64/7zz
… Cs7z.Service.MacOsx/Source/osx/x64/7zz
… Cs7z.Service.Windows/Cs7z.Service.Windows.csproj
… Cs7z.Service.Windows/Source/Windows/arm64/7-ZipFar.dll
… Cs7z.Service.Windows/Source/Windows/arm64/7za.dll
… Cs7z.Service.Windows/Source/Windows/arm64/7za.exe
… Cs7z.Service.Windows/Source/Windows/arm64/7zxa.dll
… Cs7z.Service.Windows/Source/Windows/x64/7za.dll
… Cs7z.Service.Windows/Source/Windows/x64/7za.exe
… Cs7z.Service.Windows/Source/Windows/x64/7zxa.dll
… Cs7z.Service.Windows/WindowsSevenZipExecutableSource.cs
… Cs7z.sln