## Summary

This PR adds comprehensive multi-architecture support, compression level control, and performance benchmarking capabilities to the Cs7z library.

## Key Features

### 🏗️ Multi-Architecture Support
- **Windows**: x86, x64, ARM64
- **Linux**: x86, x64, ARM, ARM64  
- **macOS**: x64, ARM64
- OmniPlatform service automatically detects OS and architecture
- Bundled 7-zip executables for all supported platforms

### 🎚️ Compression Level Support
- Added `CompressionLevel` enum with 6 levels:
  - Store (0) - No compression
  - Fastest (1) - Minimal compression
  - Fast (3) - Quick compression
  - Normal (5) - Balanced (default)
  - Maximum (7) - High compression
  - Ultra (9) - Best compression
- New overload: `CreateArchive(path, folder, compressionLevel)`
- Backward compatible - existing code continues to work

### 📊 Performance Benchmarking
- Comprehensive performance console comparing:
  - 7-Zip vs .NET ZipArchive
  - Different compression levels
  - Various file sizes (1MB to 1GB)
- BenchmarkDotNet integration for detailed analysis

## Changes

- ✨ Added `CompressionLevel` enum in `Cs7z.Core.Models`
- ✨ Updated `ISevenZipArchive` interface with compression level overload
- ✨ Modified `SevenZipArchive` to support `-mx` parameter
- ✨ Added multi-architecture executable detection
- 🧪 Added comprehensive unit tests for compression levels
- 🧪 Added integration tests verifying compression ratios
- 🧪 Created performance benchmarking projects
- 📝 Updated CLAUDE.md with current project structure
- 🐳 Added Dockerfile for cross-platform testing

## Performance Results

From benchmarking 1GB of mixed data:

**Compression Speed (1GB file):**
- Store: 322ms (no compression)
- Fastest: 4,002ms (27.9% compression)
- Normal: 27,018ms (33.7% compression)  
- Ultra: 43,835ms (33.7% compression)

**Key Findings:**
- Ultra is only 0.1% better than Normal but 62% slower
- 7-Zip achieves 5-6% better compression than .NET ZipArchive
- When targeting similar compression ratios, 7-Zip is actually faster

## Testing

- ✅ All unit tests pass
- ✅ All integration tests pass
- ✅ Performance benchmarks complete
- ✅ Cross-platform testing via Docker
- ✅ Backward compatibility verified

## Breaking Changes

None - all changes are additive and backward compatible.

## Usage Example

```csharp
// Using compression levels
var archive = new SevenZipArchive(new OmniPlatformSevenZipExecutableSource());

// Default (Normal compression)
await archive.CreateArchive("output.7z", "sourceFolder");

// Fast compression
await archive.CreateArchive("output.7z", "sourceFolder", CompressionLevel.Fastest);

// Maximum compression
await archive.CreateArchive("output.7z", "sourceFolder", CompressionLevel.Ultra);
```

🤖 Generated with [Claude Code](https://claude.ai/code)